<!–- Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at

http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. –>

Arrow.jl

User Manual

Support for generic path-like types

Reading arrow data

Arrow.Table

Arrow types

Arrow.Stream

Custom application metadata

Writing arrow data

Arrow.write

Arrow.Writer

Multithreaded writing

Compression

API Reference

Arrow — Module

Arrow.jl

A pure Julia implementation of the apache arrow memory format specification.

This implementation supports the 1.0 version of the specification, including support for:

All primitive data types

All nested data types

Dictionary encodings, nested dictionary encodings, and messages

Extension types

Streaming, file, record batch, and replacement and isdelta dictionary messages

Buffer compression/decompression via the standard LZ4 frame and Zstd formats

It currently doesn't include support for:

Tensors or sparse tensors

Flight RPC

C data interface

Third-party data formats:

csv and parquet support via the existing CSV.jl and Parquet.jl packages

Other Tables.jl-compatible packages automatically supported (DataFrames.jl, JSONTables.jl, JuliaDB.jl, SQLite.jl, MySQL.jl, JDBC.jl, ODBC.jl, XLSX.jl, etc.)

No current Julia packages support ORC or Avro data formats

See docs for official Arrow.jl API with the User Manual and reference docs for Arrow.Table, Arrow.write, and Arrow.Stream.